@@ -9,8 +9,10 @@
   "scripts": {
     "start": "node devserver.js",
     "build:clean": "rimraf ./public/dist",
-    "build:prod": "npm run build:clean && cross-env NODE_ENV=production webpack --config webpack.config.prod.js",
-    "build:prod:withstats": "npm run build:clean && cross-env NODE_ENV=production webpack --config webpack.config.prod.js --profile --json > webpack-stats.json",
+    "build:prod":
+      "npm run build:clean && cross-env NODE_ENV=production webpack --config webpack.config.prod.js",
+    "build:prod:withstats":
+      "npm run build:clean && cross-env NODE_ENV=production webpack --config webpack.config.prod.js --profile --json > webpack-stats.json",
     "test": "jest --coverage",
     "watch:test": "jest --watch",
     "manage:translations": "babel-node ./translationRunner.js",
@@ -19,43 +21,6 @@
   },
   "author": "Jannik Hell",
   "license": "MIT",
-  "//": [
-    "________________________devDependencies________________________",
-    "autoprefixer: used in webpack post-css loader to prefix css verbs with vendor specific css verbs",
-    "babel-cli: includes babel-node executable which is used to npm run manage:translations",
-    "babel-core: includes the babel compiler which is used to transform es6 to es5 code",
-    "babel-jest: will use babel to test es6 code with jest",
-    "babel-loader",
-    "babel-plugin-react-intl",
-    "babel-preset-es2015",
-    "babel-preset-react",
-    "babel-preset-stage-0",
-    "cross-env",
-    "css-loader",
-    "eslint",
-    "eslint-plugin-react",
-    "extract-text-webpack-plugin",
-    "jest-cli",
-    "node-sass",
-    "postcss-loader",
-    "react-hot-loader",
-    "react-intl-translations-manager",
-    "rimraf",
-    "sass-loader",
-    "style-loader",
-    "webpack",
-    "webpack-dev-server",
-    "webpack-validator: validates webpack config within config file with validate(config)",
-    "________________________dependencies________________________",
-    "json-loader",
-    "moment: consistent date and time API with types, formatting, modifying",
-    "react",
-    "react-dom",
-    "react-intl: provides language specific string, date/time and number formatting",
-    "react-redux: provides redux store to react components with connect(component)",
-    "redux: flux state management (inspired by Elm architecture)",
-    "redux-thunk: redux middleware to use asynchronous actions"
-  ],
   "devDependencies": {
     "@kadira/react-storybook-addon-info": "^3.3.0",
     "@kadira/storybook": "^2.21.0",
@@ -78,12 +43,12 @@
     "jest-cli": "^19.0.2",
     "node-sass": "^4.5.0",
     "postcss-loader": "^1.3.3",
-    "react-hot-loader": "^1.3.1",
+    "react-hot-loader": "^3.0.0-beta.7",
     "react-intl-translations-manager": "^4.0.1",
     "rimraf": "^2.6.1",
     "sass-loader": "^6.0.3",
     "style-loader": "^0.13.2",
-    "webpack": "^2.2.1",
+    "webpack": "3.0.0",
     "webpack-dev-server": "^2.4.1",
     "webpack-validator": "^2.3.0"
   },
@@ -91,8 +56,9 @@
     "intl": "^1.2.5",
     "json-loader": "^0.5.4",
     "moment": "^2.17.1",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "prop-types": "^15.6.0",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "react-intl": "^2.2.3",
     "react-redux": "^5.0.3",
     "redux": "^3.6.0",
@@ -100,8 +66,6 @@
   },
   "jest": {
     "testEnvironment": "node",
-    "collectCoverageFrom": [
-      "src/client/**/*.js"
-    ]
+    "collectCoverageFrom": ["src/client/**/*.js"]
   }
 }
