@@ -1,24 +1,25 @@
-import React from 'react';
-import { IntlProvider } from 'react-intl';
+import PropTypes from 'prop-types';
+import React from 'react'
+import { IntlProvider } from 'react-intl'
 
 // Import Redux Store Provider Component
-import { Provider } from 'react-redux';
+import { Provider } from 'react-redux'
 
 // Import our Component
-import HelloWorld from '../components/HelloWorld.js';
+import HelloWorld from '../components/HelloWorld.js'
 
 export default class Root extends React.Component {
   constructor(props) {
-    super(props);
+    super(props)
 
     this.state = {
       language: 'de'
-    };
+    }
   }
 
   render() {
-    const { language } = this.state;
-    const messages = require('../translations/locales/' + language + '.json');
+    const { language } = this.state
+    const messages = require('../translations/locales/' + language + '.json')
 
     return (
       <IntlProvider locale={language} key={language} messages={messages}>
@@ -37,10 +38,10 @@ export default class Root extends React.Component {
           </div>
         </Provider>
       </IntlProvider>
-    );
+    )
   }
 }
 
 Root.propTypes = {
-  store: React.PropTypes.object.isRequired
-};
+  store: PropTypes.object.isRequired
+}
